rename discount index to tsx
